+++ madeasy/airline/urls.py
@@ -7,6 +7,8 @@ from .views import (
     TravelClassViewSet,
     FlightViewSet,
     TravelClassSeatCapacityViewSet,
+    RouteTreeViewSet,
+    RouteViewSet
 )
 
 
@@ -16,6 +18,9 @@ router.register(r'airlines', AirlineViewSet)
 router.register(r'airplanes', AirplaneViewSet)
 router.register(r'travel_classes', TravelClassViewSet)
 router.register(r'flights', FlightViewSet)
+router.register(r'flights', FlightViewSet)
+router.register(r'routes', RouteViewSet)
+router.register(r'route_trees', RouteTreeViewSet)
 router.register(
     r'travel_class_seat_capacities', TravelClassSeatCapacityViewSet)